+++ Script/src/main/org.l2j.scripts/org/l2j/scripts/handlers/effecthandlers/Mp.java
@@ -58,7 +58,7 @@ public final class Mp extends AbstractEffect {
             return;
         }
 
-        int basicAmount = power;
+        double basicAmount = power;
         if ((item != null) && (item.isPotion() || item.isElixir())) {
             basicAmount += effected.getStats().getValue(Stat.ADDITIONAL_POTION_MP, 0);
         }
